chore: add vertical lines to the explorer
